Aplicação/backend/requirements.txt
# Core FastAPI Stack
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
SQLAlchemy[asyncio]==2.0.43
asyncpg==0.30.0
alembic==1.16.5
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.2
reportlab==4.2.5
python-jose[cryptography]==3.3.0
email-validator==2.2.0
google-generativeai==0.8.3

# Development & Testing
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
aiosqlite==0.21.0

# Code Quality
ruff==0.8.4
black==25.1.0
mypy==1.17.1

# Additional utilities
python-dotenv==1.1.1
structlog==24.4.0
prometheus-client==0.22.1